Refactor fc-agent: async Rust, no globals, clean module structure
Split monolithic main.rs (3317 lines) into 14 focused modules:

  main.rs      40 lines   Entry point + tracing init
  agent.rs    176 lines   Orchestrator
  types.rs    154 lines   Plan, VolumeMount, ExtraDiskMount, NfsMount, LatestMetadata
  vsock.rs    179 lines   VsockStream (AsyncFd), VsockListener, send_status
  output.rs   149 lines   OutputHandle (async mpsc), writer task
  mmds.rs     227 lines   MMDS plan fetch, clock sync, restore-epoch watcher
  network.rs  340 lines   DNS, IPv6, ARP, TCP kill, localhost forwarding
  exec.rs     212 lines   Exec server using VsockListener
  restore.rs  268 lines   Snapshot restore, FUSE remount, cross-ns rebind
  mounts.rs   275 lines   FUSE/disk/NFS mounting + cleanup
  container.rs 528 lines  Image import/pull, podman args, run (TTY + async)
  system.rs   326 lines   Cgroups, KVM, rlimits, shutdown, proxy
  lock_test.rs 109 lines  File locking tests for clone validation

Design improvements:
- Eliminated global OUTPUT_VSOCK_FD AtomicI32 — replaced with OutputHandle
  backed by tokio::sync::mpsc channel. No atomics, no raw fd management
  in business logic.
- Async vsock I/O via VsockStream wrapping AsyncFd<OwnedFd>. No dedicated
  writer thread — writes happen in tokio tasks.
- Reconnect via channel message (OutputMessage::Reconnect) instead of
  atomic fd swap race.
- nix crate for socket operations instead of raw libc. OwnedFd for RAII
  fd management — no manual libc::close.
- VsockListener using nix + AsyncFd for exec server accept loop.

8 unit tests: Plan deserialization, OutputHandle send/reconnect/shutdown.
Sanity test (rootless) passes with the refactored fc-agent.
