+++ fc-agent/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 anyhow = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "process", "fs", "io-util", "signal"] }
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "process", "fs", "io-util", "signal", "sync", "time"] }
 reqwest = { version = "0.11", default-features = false, features = ["json"] }
 libc = "0.2"
 nix = { version = "0.29", features = ["socket", "poll", "fs"] }
